test: 중복 테스트 파일 통합
